+++ arrow-schema/src/schema.rs
@@ -286,6 +286,8 @@ impl Hash for Schema {
 mod tests {
     use super::*;
     use crate::datatype::DataType;
+    use crate::{TimeUnit, UnionMode};
+    use std::collections::BTreeMap;
 
     #[test]
     #[cfg(feature = "serde")]
@@ -382,5 +384,399 @@ mod tests {
 
         assert!(!schema1.contains(&schema2));
         assert!(schema2.contains(&schema1));
+    }
+
+    #[test]
+    fn schema_equality() {
+        let schema1 = Schema::new(vec![
+            Field::new("c1", DataType::Utf8, false),
+            Field::new("c2", DataType::Float64, true),
+            Field::new("c3", DataType::LargeBinary, true),
+        ]);
+        let schema2 = Schema::new(vec![
+            Field::new("c1", DataType::Utf8, false),
+            Field::new("c2", DataType::Float64, true),
+            Field::new("c3", DataType::LargeBinary, true),
+        ]);
+
+        assert_eq!(schema1, schema2);
+
+        let schema3 = Schema::new(vec![
+            Field::new("c1", DataType::Utf8, false),
+            Field::new("c2", DataType::Float32, true),
+        ]);
+        let schema4 = Schema::new(vec![
+            Field::new("C1", DataType::Utf8, false),
+            Field::new("C2", DataType::Float64, true),
+        ]);
+
+        assert_ne!(schema1, schema3);
+        assert_ne!(schema1, schema4);
+        assert_ne!(schema2, schema3);
+        assert_ne!(schema2, schema4);
+        assert_ne!(schema3, schema4);
+
+        let f = Field::new("c1", DataType::Utf8, false).with_metadata(Some(
+            [("foo".to_string(), "bar".to_string())]
+                .iter()
+                .cloned()
+                .collect(),
+        ));
+        let schema5 = Schema::new(vec![
+            f,
+            Field::new("c2", DataType::Float64, true),
+            Field::new("c3", DataType::LargeBinary, true),
+        ]);
+        assert_ne!(schema1, schema5);
+    }
+
+    #[test]
+    fn create_schema_string() {
+        let schema = person_schema();
+        assert_eq!(schema.to_string(),
+                   "Field { name: \"first_name\", data_type: Utf8, nullable: false, dict_id: 0, dict_is_ordered: false, metadata: Some({\"k\": \"v\"}) }, \
+        Field { name: \"last_name\", data_type: Utf8, nullable: false, dict_id: 0, dict_is_ordered: false, metadata: None }, \
+        Field { name: \"address\", data_type: Struct([\
+            Field { name: \"street\", data_type: Utf8, nullable: false, dict_id: 0, dict_is_ordered: false, metadata: None }, \
+            Field { name: \"zip\", data_type: UInt16, nullable: false, dict_id: 0, dict_is_ordered: false, metadata: None }\
+        ]), nullable: false, dict_id: 0, dict_is_ordered: false, metadata: None }, \
+        Field { name: \"interests\", data_type: Dictionary(Int32, Utf8), nullable: true, dict_id: 123, dict_is_ordered: true, metadata: None }")
+    }
+
+    #[test]
+    fn schema_field_accessors() {
+        let schema = person_schema();
+
+        // test schema accessors
+        assert_eq!(schema.fields().len(), 4);
+
+        // test field accessors
+        let first_name = &schema.fields()[0];
+        assert_eq!(first_name.name(), "first_name");
+        assert_eq!(first_name.data_type(), &DataType::Utf8);
+        assert!(!first_name.is_nullable());
+        assert_eq!(first_name.dict_id(), None);
+        assert_eq!(first_name.dict_is_ordered(), None);
+
+        let metadata = first_name.metadata();
+        assert!(metadata.is_some());
+        let md = metadata.as_ref().unwrap();
+        assert_eq!(md.len(), 1);
+        let key = md.get("k");
+        assert!(key.is_some());
+        assert_eq!(key.unwrap(), "v");
+
+        let interests = &schema.fields()[3];
+        assert_eq!(interests.name(), "interests");
+        assert_eq!(
+            interests.data_type(),
+            &DataType::Dictionary(Box::new(DataType::Int32), Box::new(DataType::Utf8))
+        );
+        assert_eq!(interests.dict_id(), Some(123));
+        assert_eq!(interests.dict_is_ordered(), Some(true));
+    }
+
+    #[test]
+    #[should_panic(
+        expected = "Unable to get field named \\\"nickname\\\". Valid fields: [\\\"first_name\\\", \\\"last_name\\\", \\\"address\\\", \\\"interests\\\"]"
+    )]
+    fn schema_index_of() {
+        let schema = person_schema();
+        assert_eq!(schema.index_of("first_name").unwrap(), 0);
+        assert_eq!(schema.index_of("last_name").unwrap(), 1);
+        schema.index_of("nickname").unwrap();
+    }
+
+    #[test]
+    #[should_panic(
+        expected = "Unable to get field named \\\"nickname\\\". Valid fields: [\\\"first_name\\\", \\\"last_name\\\", \\\"address\\\", \\\"interests\\\"]"
+    )]
+    fn schema_field_with_name() {
+        let schema = person_schema();
+        assert_eq!(
+            schema.field_with_name("first_name").unwrap().name(),
+            "first_name"
+        );
+        assert_eq!(
+            schema.field_with_name("last_name").unwrap().name(),
+            "last_name"
+        );
+        schema.field_with_name("nickname").unwrap();
+    }
+
+    #[test]
+    fn schema_field_with_dict_id() {
+        let schema = person_schema();
+
+        let fields_dict_123: Vec<_> = schema
+            .fields_with_dict_id(123)
+            .iter()
+            .map(|f| f.name())
+            .collect();
+        assert_eq!(fields_dict_123, vec!["interests"]);
+
+        assert!(schema.fields_with_dict_id(456).is_empty());
+    }
+
+    fn person_schema() -> Schema {
+        let kv_array = [("k".to_string(), "v".to_string())];
+        let field_metadata: BTreeMap<String, String> = kv_array.iter().cloned().collect();
+        let first_name = Field::new("first_name", DataType::Utf8, false)
+            .with_metadata(Some(field_metadata));
+
+        Schema::new(vec![
+            first_name,
+            Field::new("last_name", DataType::Utf8, false),
+            Field::new(
+                "address",
+                DataType::Struct(vec![
+                    Field::new("street", DataType::Utf8, false),
+                    Field::new("zip", DataType::UInt16, false),
+                ]),
+                false,
+            ),
+            Field::new_dict(
+                "interests",
+                DataType::Dictionary(Box::new(DataType::Int32), Box::new(DataType::Utf8)),
+                true,
+                123,
+                true,
+            ),
+        ])
+    }
+
+    #[test]
+    fn test_try_merge_field_with_metadata() {
+        // 1. Different values for the same key should cause error.
+        let metadata1: BTreeMap<String, String> =
+            [("foo".to_string(), "bar".to_string())]
+                .iter()
+                .cloned()
+                .collect();
+        let f1 = Field::new("first_name", DataType::Utf8, false)
+            .with_metadata(Some(metadata1));
+
+        let metadata2: BTreeMap<String, String> =
+            [("foo".to_string(), "baz".to_string())]
+                .iter()
+                .cloned()
+                .collect();
+        let f2 = Field::new("first_name", DataType::Utf8, false)
+            .with_metadata(Some(metadata2));
+
+        assert!(
+            Schema::try_merge(vec![Schema::new(vec![f1]), Schema::new(vec![f2])])
+                .is_err()
+        );
+
+        // 2. None + Some
+        let mut f1 = Field::new("first_name", DataType::Utf8, false);
+        let metadata2: BTreeMap<String, String> =
+            [("missing".to_string(), "value".to_string())]
+                .iter()
+                .cloned()
+                .collect();
+        let f2 = Field::new("first_name", DataType::Utf8, false)
+            .with_metadata(Some(metadata2));
+
+        assert!(f1.try_merge(&f2).is_ok());
+        assert!(f1.metadata().is_some());
+        assert_eq!(
+            f1.metadata().as_ref().unwrap(),
+            f2.metadata().as_ref().unwrap()
+        );
+
+        // 3. Some + Some
+        let mut f1 = Field::new("first_name", DataType::Utf8, false).with_metadata(Some(
+            [("foo".to_string(), "bar".to_string())]
+                .iter()
+                .cloned()
+                .collect(),
+        ));
+        let f2 = Field::new("first_name", DataType::Utf8, false).with_metadata(Some(
+            [("foo2".to_string(), "bar2".to_string())]
+                .iter()
+                .cloned()
+                .collect(),
+        ));
+
+        assert!(f1.try_merge(&f2).is_ok());
+        assert!(f1.metadata().is_some());
+        assert_eq!(
+            f1.metadata().cloned().unwrap(),
+            [
+                ("foo".to_string(), "bar".to_string()),
+                ("foo2".to_string(), "bar2".to_string())
+            ]
+            .iter()
+            .cloned()
+            .collect()
+        );
+
+        // 4. Some + None.
+        let mut f1 = Field::new("first_name", DataType::Utf8, false).with_metadata(Some(
+            [("foo".to_string(), "bar".to_string())]
+                .iter()
+                .cloned()
+                .collect(),
+        ));
+        let f2 = Field::new("first_name", DataType::Utf8, false);
+        assert!(f1.try_merge(&f2).is_ok());
+        assert!(f1.metadata().is_some());
+        assert_eq!(
+            f1.metadata().cloned().unwrap(),
+            [("foo".to_string(), "bar".to_string())]
+                .iter()
+                .cloned()
+                .collect()
+        );
+
+        // 5. None + None.
+        let mut f1 = Field::new("first_name", DataType::Utf8, false);
+        let f2 = Field::new("first_name", DataType::Utf8, false);
+        assert!(f1.try_merge(&f2).is_ok());
+        assert!(f1.metadata().is_none());
+    }
+
+    #[test]
+    fn test_schema_merge() {
+        let merged = Schema::try_merge(vec![
+            Schema::new(vec![
+                Field::new("first_name", DataType::Utf8, false),
+                Field::new("last_name", DataType::Utf8, false),
+                Field::new(
+                    "address",
+                    DataType::Struct(vec![Field::new("zip", DataType::UInt16, false)]),
+                    false,
+                ),
+            ]),
+            Schema::new_with_metadata(
+                vec![
+                    // nullable merge
+                    Field::new("last_name", DataType::Utf8, true),
+                    Field::new(
+                        "address",
+                        DataType::Struct(vec![
+                            // add new nested field
+                            Field::new("street", DataType::Utf8, false),
+                            // nullable merge on nested field
+                            Field::new("zip", DataType::UInt16, true),
+                        ]),
+                        false,
+                    ),
+                    // new field
+                    Field::new("number", DataType::Utf8, true),
+                ],
+                [("foo".to_string(), "bar".to_string())]
+                    .iter()
+                    .cloned()
+                    .collect::<HashMap<String, String>>(),
+            ),
+        ])
+        .unwrap();
+
+        assert_eq!(
+            merged,
+            Schema::new_with_metadata(
+                vec![
+                    Field::new("first_name", DataType::Utf8, false),
+                    Field::new("last_name", DataType::Utf8, true),
+                    Field::new(
+                        "address",
+                        DataType::Struct(vec![
+                            Field::new("zip", DataType::UInt16, true),
+                            Field::new("street", DataType::Utf8, false),
+                        ]),
+                        false,
+                    ),
+                    Field::new("number", DataType::Utf8, true),
+                ],
+                [("foo".to_string(), "bar".to_string())]
+                    .iter()
+                    .cloned()
+                    .collect::<HashMap<String, String>>()
+            )
+        );
+
+        // support merge union fields
+        assert_eq!(
+            Schema::try_merge(vec![
+                Schema::new(vec![Field::new(
+                    "c1",
+                    DataType::Union(
+                        vec![
+                            Field::new("c11", DataType::Utf8, true),
+                            Field::new("c12", DataType::Utf8, true),
+                        ],
+                        vec![0, 1],
+                        UnionMode::Dense
+                    ),
+                    false
+                ),]),
+                Schema::new(vec![Field::new(
+                    "c1",
+                    DataType::Union(
+                        vec![
+                            Field::new("c12", DataType::Utf8, true),
+                            Field::new("c13", DataType::Time64(TimeUnit::Second), true),
+                        ],
+                        vec![1, 2],
+                        UnionMode::Dense
+                    ),
+                    false
+                ),])
+            ])
+            .unwrap(),
+            Schema::new(vec![Field::new(
+                "c1",
+                DataType::Union(
+                    vec![
+                        Field::new("c11", DataType::Utf8, true),
+                        Field::new("c12", DataType::Utf8, true),
+                        Field::new("c13", DataType::Time64(TimeUnit::Second), true),
+                    ],
+                    vec![0, 1, 2],
+                    UnionMode::Dense
+                ),
+                false
+            ),]),
+        );
+
+        // incompatible field should throw error
+        assert!(Schema::try_merge(vec![
+            Schema::new(vec![
+                Field::new("first_name", DataType::Utf8, false),
+                Field::new("last_name", DataType::Utf8, false),
+            ]),
+            Schema::new(vec![Field::new("last_name", DataType::Int64, false),])
+        ])
+        .is_err());
+
+        // incompatible metadata should throw error
+        let res = Schema::try_merge(vec![
+            Schema::new_with_metadata(
+                vec![Field::new("first_name", DataType::Utf8, false)],
+                [("foo".to_string(), "bar".to_string())]
+                    .iter()
+                    .cloned()
+                    .collect::<HashMap<String, String>>(),
+            ),
+            Schema::new_with_metadata(
+                vec![Field::new("last_name", DataType::Utf8, false)],
+                [("foo".to_string(), "baz".to_string())]
+                    .iter()
+                    .cloned()
+                    .collect::<HashMap<String, String>>(),
+            ),
+        ])
+        .unwrap_err();
+
+        let expected = "Fail to merge schema due to conflicting metadata. Key 'foo' has different values 'bar' and 'baz'";
+        assert!(
+            res.to_string().contains(expected),
+            "Could not find expected string '{}' in '{}'",
+            expected,
+            res
+        );
     }
 }
